refactor: extract buildRegistry/newMux so real wiring is tested
Close two test-fidelity gaps from the post-v3 review without changing any
behavior:

- Extract buildRegistry (cmd/exporter/main.go) from run(). The build-info
  test now calls it directly instead of re-implementing the WrapRegistererWith
  wiring, so deleting the tdarr_instance-labeled build_info registration now
  fails a test (previously it failed none).

- Extract newMux (internal/server/server.go) — the routes, catch-all, and
  Recovery/RequestLogger stack — from ServeHttp. The mux integration tests move
  from internal/handlers/handlers_test.go into internal/server/server_test.go
  and exercise the real newMux, so routing/middleware drift in server.go is now
  caught. handlers_test.go's duplicated handler stack is removed; package
  handlers keeps middleware_test.go.
